+++ mundial-ticketing/mundial-ticketing/backend/src/main/java/com/ucu/ticketing/dto/response/EstadioResponse.java
@@ -6,13 +6,12 @@ import lombok.NoArgsConstructor;
 
 import java.util.List;
 
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class EstadioResponse {
 
-    private Integer idEstadio;
+    private Long idEstadio;
     private String nombre;
     private String nombrePais;
     private List<SectorResponse> sectores;
