+++ mundial-ticketing/mundial-ticketing/backend/src/main/java/com/ucu/ticketing/dto/response/ValidacionResponse.java
@@ -14,7 +14,7 @@ public class ValidacionResponse {
 
     private boolean accesoPermitido;
     private String mensaje;
-    private Integer idEntrada;
+    private Long idEntrada;
     private String letraSector;
     private String mailPropietario;
     private LocalDateTime horaIngreso;
